Add Computing as a stand-alone subject
38 candidates have entered 'Computing' as their degree subject.

Whilst "Computing and information technology" is in the list, as does "Information technology", "Creative computing", "Applied computing", "Parallel computing", "Neural computing", "High performance computing", "Business computing", "Computer science", and so on... - "Computing" is not listed as a standalone subject.

Examples of undergraduate courses which use 'Computing' as degree title:

* https://www.gre.ac.uk/undergraduate-courses/ach/computing-bsc-hons
* https://www.imperial.ac.uk/study/ug/courses/computing-department/computing-beng/
* https://www.port.ac.uk/study/courses/bsc-hons-computing

The HESA code 100367 (Computing and information technology) is reused as this is likely the closest match.